+++ packages/blade/src/components/Tabs/TabsItem.tsx
@@ -6,14 +6,12 @@ import { CompositeItem } from '@floating-ui/react';
 import type { TabsItemProps, TabsProps } from './types';
 import { useTabsContext } from './TabsContext';
 import { backgroundColor, paddings, textColor } from './tabTokens';
+import { iconSizeMap, useTabsItemPropRestriction } from './utils';
 import { Text } from '~components/Typography';
 import { castWebType, makeBorderSize, makeMotionTime, makeSpace } from '~utils';
 import useInteraction from '~utils/useInteraction';
 import { useIsMobile } from '~utils/useIsMobile';
-import { logger, throwBladeError } from '~utils/logger';
-import { getComponentId } from '~utils/isValidAllowedChildren';
 import { makeAccessible } from '~utils/makeAccessible';
-import { useIsomorphicLayoutEffect } from '~utils/useIsomorphicLayoutEffect';
 
 const StyledTabButton = styled.button<{
   size: TabsProps['size'];
@@ -23,37 +21,35 @@ const StyledTabButton = styled.button<{
   isSelected: boolean;
 }>(({ theme, isSelected, size, variant, autoWidth, isVertical }) => {
   const isMobile = useIsMobile();
+  const isFilled = variant === 'filled';
   const device = isMobile ? 'mobile' : 'desktop';
   const orientation = isVertical ? 'vertical' : 'horizontal';
+  const border = isVertical ? 'borderLeft' : 'borderBottom';
   const selectedState = isSelected ? 'selected' : 'unselected';
-  const isFilled = variant === 'filled';
+  const background = backgroundColor[selectedState][variant];
 
   const getColor = (value: string): string => {
     if (value === 'transparent') return 'transparent';
     return get(theme, value);
   };
 
-  const border = isVertical ? 'borderLeft' : 'borderBottom';
-
   return {
-    width: autoWidth ? '100%' : undefined,
     appearance: 'none',
     border: 'none',
     outline: 'none',
     cursor: 'pointer',
     display: 'flex',
     alignItems: 'center',
-    justifyContent: autoWidth && !isVertical ? 'center' : 'left',
+    justifyContent: isVertical ? 'left' : 'center',
     gap: makeSpace(theme.spacing[3]),
+    width: autoWidth ? '100%' : undefined,
     paddingTop: makeSpace(get(theme, paddings[variant][orientation][device].top[size!])),
     paddingBottom: makeSpace(get(theme, paddings[variant][orientation][device].bottom[size!])),
     paddingLeft: makeSpace(get(theme, paddings[variant][orientation][device].left[size!])),
     paddingRight: makeSpace(get(theme, paddings[variant][orientation][device].right[size!])),
     // colors
     backgroundColor:
-      isSelected && isFilled && !isVertical
-        ? 'transparent'
-        : getColor(backgroundColor[selectedState][variant].default),
+      isSelected && isFilled && !isVertical ? 'transparent' : getColor(background.default),
     borderRadius: isFilled && !isVertical ? theme.border.radius.small : 0,
     [`${border}Style`]: 'solid',
     [`${border}Width`]: isFilled ? 0 : makeBorderSize(theme.border.width.thick),
@@ -68,18 +64,16 @@ const StyledTabButton = styled.button<{
       backgroundColor:
         // Don't want to show hover state on filled tabs when vertical because
         // The hover color needs to be on the TabIndicator instead.
-        isSelected && isFilled && !isVertical
-          ? 'transparent'
-          : getColor(backgroundColor[selectedState][variant].hover),
+        isSelected && isFilled && !isVertical ? 'transparent' : getColor(background.hover),
     },
     '&:disabled': {
       cursor: 'not-allowed',
-      backgroundColor: getColor(backgroundColor[selectedState][variant].disabled),
+      backgroundColor: getColor(background.disabled),
     },
     '&:focus-visible': {
       borderRadius: makeSpace(theme.border.radius.medium),
       boxShadow: `0px 0px 0px 4px ${theme.colors.brand.primary[400]}`,
-      backgroundColor: getColor(backgroundColor[selectedState][variant].focus),
+      backgroundColor: getColor(background.focus),
     },
 
     transitionProperty: 'all',
@@ -92,84 +86,6 @@ const StyledTabButton = styled.button<{
     },
   };
 });
-
-const iconSizeMap = {
-  medium: 'medium',
-  large: 'large',
-} as const;
-
-const badgeSizeMap = {
-  medium: 'small',
-  large: 'medium',
-} as const;
-
-const counterSizeMap = {
-  medium: 'small',
-  large: 'small',
-} as const;
-
-const propRestrictionMap = {
-  Badge: {
-    medium: {
-      size: badgeSizeMap.medium,
-    },
-    large: {
-      size: badgeSizeMap.large,
-    },
-  },
-  Counter: {
-    medium: {
-      size: counterSizeMap.medium,
-    },
-    large: {
-      size: counterSizeMap.large,
-    },
-  },
-} as const;
-
-type TrailingComponents = keyof typeof propRestrictionMap;
-
-const useTrailingRestriction = (
-  trailing: React.ReactNode,
-  tabItemSize: NonNullable<TabsProps['size']>,
-): React.ReactNode => {
-  const [
-    validatedTrailingComponent,
-    setValidatedTrailingComponent,
-  ] = React.useState<React.ReactElement | null>(null);
-
-  // validate and restrict sub component props in trailing prop
-  useIsomorphicLayoutEffect(() => {
-    if (React.isValidElement(trailing)) {
-      const trailingComponentType = getComponentId(trailing) as TrailingComponents;
-      const restrictedProps = propRestrictionMap[trailingComponentType]?.[tabItemSize];
-      if (__DEV__) {
-        if (!restrictedProps) {
-          throwBladeError({
-            message: `Only Badge or Counter component is accepted as trailing`,
-            moduleName: 'TabsItem',
-          });
-        }
-
-        const restrictedPropKeys = Object.keys(restrictedProps);
-        for (const prop of restrictedPropKeys) {
-          if (trailing?.props?.hasOwnProperty(prop)) {
-            logger({
-              message: `Do not pass "${prop}" to "${trailingComponentType}" while inside TabsItem trailing, because we override it.`,
-              moduleName: 'TabsItem',
-              type: 'warn',
-            });
-          }
-        }
-      }
-      setValidatedTrailingComponent(
-        React.cloneElement(trailing as React.ReactElement, restrictedProps),
-      );
-    }
-  }, [tabItemSize, trailing]);
-
-  return validatedTrailingComponent;
-};
 
 const TabsItem = ({
   children,
@@ -188,7 +104,7 @@ const TabsItem = ({
     isVertical,
   } = useTabsContext();
   const { currentInteraction, ...interactionProps } = useInteraction();
-  const validatedTrailingComponent = useTrailingRestriction(trailing, size!);
+  const validatedTrailingComponent = useTabsItemPropRestriction(trailing, size!);
   const isSelected = selectedValue === value;
   const selectedState = isSelected ? 'selected' : 'unselected';
   const panelId = `${baseId}-${value}-tabpanel`;
